Fixed bug
When clicking filter and then clicking outside the box, the filter icon
would change to white.
So we would have a white background and white icon. Which made it
difficult to see the icon.

The icon colour should change depending on whether a filter is
selected. This solves that issue.